+++ src/test/java/racingcar/GameManagerTest.java
@@ -18,30 +18,4 @@ public class GameManagerTest {
         GameManager gameManager = new GameManager(user);
         System.out.println(gameManager.announceWinner());
     }
-
-    private class GameManager {
-        private User user;
-        private List<String> winnerMembers = new ArrayList<>();
-
-        public GameManager(User user) {
-            this.user = user;
-        }
-        public String announceWinner() {
-            int longestDistanceDriven = -1;
-
-            for (Car car : user.getCars()) {
-                if (longestDistanceDriven < car.getDistanceDriven().length()) {
-                    longestDistanceDriven = car.getDistanceDriven().length();
-                }
-            }
-
-            for (Car car : user.getCars()) {
-                if (longestDistanceDriven == car.getDistanceDriven().length()) {
-                    winnerMembers.add(car.getCarName());
-                }
-            }
-
-            return String.join(", ", winnerMembers);
-        }
-    }
 }
